requirements.txt
google-api-python-client>=2.140.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
google-auth>=2.34.0
tqdm>=4.66.0
pyinstaller>=6.0.0
PyQt5>=5.15.0